refactor: application -> domain으로 위치 변경
