+++ src/main/kotlin/mine/sweeper/domain/Field.kt
@@ -1,6 +1,6 @@
 package mine.sweeper.domain
 
-import mine.sweeper.application.value.GameStatus
+import mine.sweeper.domain.value.GameStatus
 import mine.sweeper.view.dto.Position
 
 sealed class Field(val position: Position) {
