+++ src/main/kotlin/mine/sweeper/domain/value/Height.kt
@@ -1,4 +1,4 @@
-package mine.sweeper.application.value
+package mine.sweeper.domain.value
 
 @JvmInline
 value class Height(val value: Int) {
