+++ src/main/kotlin/mine/sweeper/domain/value/MineCount.kt
@@ -1,4 +1,4 @@
-package mine.sweeper.application.value
+package mine.sweeper.domain.value
 
 @JvmInline
 value class MineCount(val value: Int) {
