+++ src/main/kotlin/mine/sweeper/view/InputView.kt
@@ -1,8 +1,8 @@
 package mine.sweeper.view
 
-import mine.sweeper.application.value.Height
-import mine.sweeper.application.value.MineCount
-import mine.sweeper.application.value.Width
+import mine.sweeper.domain.value.Height
+import mine.sweeper.domain.value.MineCount
+import mine.sweeper.domain.value.Width
 import mine.sweeper.view.dto.Position
 
 object InputView {
