+++ src/test/kotlin/mine/sweeper/MineSweeperGameTest.kt
@@ -7,10 +7,10 @@ import io.kotest.matchers.shouldBe
 import mine.sweeper.Fixture.createMineGame
 import mine.sweeper.application.MapInitializer
 import mine.sweeper.application.MineSweeperGame
-import mine.sweeper.application.value.GameStatus
-import mine.sweeper.application.value.Height
-import mine.sweeper.application.value.MineCount
-import mine.sweeper.application.value.Width
+import mine.sweeper.domain.value.GameStatus
+import mine.sweeper.domain.value.Height
+import mine.sweeper.domain.value.MineCount
+import mine.sweeper.domain.value.Width
 import mine.sweeper.view.dto.MapSize
 import mine.sweeper.view.dto.Position
