+++ src/test/kotlin/mine/sweeper/PositionManagerTest.kt
@@ -3,9 +3,9 @@ package mine.sweeper
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import mine.sweeper.application.RandomPositionManager
-import mine.sweeper.application.value.Height
-import mine.sweeper.application.value.MineCount
-import mine.sweeper.application.value.Width
+import mine.sweeper.domain.value.Height
+import mine.sweeper.domain.value.MineCount
+import mine.sweeper.domain.value.Width
 import mine.sweeper.view.dto.MapSize
 
 class PositionManagerTest : StringSpec({
