+++ src/test/kotlin/mine/sweeper/domain/VultureTest.kt
@@ -4,7 +4,7 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
 import mine.sweeper.Fixture.createMineGame
-import mine.sweeper.application.value.MineCount
+import mine.sweeper.domain.value.MineCount
 
 class VultureTest : StringSpec({
